+++ packages/fetch/src/lib.node.js
@@ -1,4 +1,4 @@
-import * as WebFetch from "./fetch.js"
+import { fetch as webFetch, Headers as WebHeaders, Request as WebRequest, Response as WebResponse } from "./fetch.js"
 
 export { ReadableStream, Blob, FormData  } from './package.js';
 // Electron-renderer should get the browser implementation instead of node
@@ -6,10 +6,10 @@ export { ReadableStream, Blob, FormData  } from './package.js';
 
 // Marking export as a DOM File object instead of custom class.
 export const fetch = /** @type {typeof globalThis.fetch} */
-  (typeof globalThis.fetch === "function" ? globalThis.fetch : WebFetch.fetch)
+  (typeof globalThis.fetch === "function" ? globalThis.fetch.bind(globalThis) : webFetch)
 
-export const Headers = globalThis.Headers || WebFetch.Headers
-export const Request = globalThis.Request || WebFetch.Request
-export const Response = global.Response || WebFetch.Response
+export const Headers = globalThis.Headers || WebHeaders
+export const Request = globalThis.Request || WebRequest
+export const Response = globalThis.Response || WebResponse
 
 export default fetch