+++ packages/fetch/src/utils/form-data.js
@@ -45,20 +45,21 @@ export const getBoundary = () => randomBytes(8).toString('hex');
  * @param {string} boundary
  */
 export async function * formDataIterator(form, boundary) {
+	const encoder = new TextEncoder();
 	for (const [name, value] of form) {
-		yield getHeader(boundary, name, value);
+		yield encoder.encode(getHeader(boundary, name, value));
 
 		if (isBlob(value)) {
 			// @ts-ignore - we know our streams implement aysnc iteration
 			yield * value.stream();
 		} else {
-			yield value;
+			yield encoder.encode(value);
 		}
 
-		yield carriage;
+		yield encoder.encode(carriage);
 	}
 
-	yield getFooter(boundary);
+	yield encoder.encode(getFooter(boundary));
 }
 
 /**
@@ -104,8 +105,10 @@ export const toFormData = async (source) => {
     const form = new FormData()
     const parts = iterateMultipart(body, boundary)
     for await (const { name, data, filename, contentType } of parts) {
-      if (filename) {
+      if (typeof filename === 'string') {
         form.append(name, new File([data], filename, { type: contentType }))
+      } else if (typeof filename !== 'undefined') {
+        form.append(name, new File([], '', { type: contentType }))
       } else {
         form.append(name, new TextDecoder().decode(data), filename)
       }